+++ src/YourFile.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Walsgit\Discussion\Cards;
+
+class YourFile
+{
+}